+++ src/main/java/dev/matheus/whostheplayer/dto/PlayerOption.java
@@ -1,15 +1,15 @@
 package dev.matheus.whostheplayer.dto;
 
 public class PlayerOption {
-    private final int playerId;
+    private final Long playerId;
     private final String playerName;
 
-    public PlayerOption(int playerId, String playerName) {
+    public PlayerOption(Long playerId, String playerName) {
         this.playerId = playerId;
         this.playerName = playerName;
     }
 
-    public int getPlayerId() {
+    public Long getPlayerId() {
         return playerId;
     }
